question-core/question-core/requirements.txt
asgiref==3.3.1
astroid==2.5.1
attrs==20.3.0
cheroot==8.5.2
CherryPy==18.6.0
coverage==5.5
Django==3.1.7
django-cors-headers==3.7.0
django-material-admin==1.7.16
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
flake8==3.8.4
freezegun==1.1.0
iniconfig==1.1.1
isort==5.8.0
jaraco.classes==3.2.1
jaraco.collections==3.2.0
jaraco.functools==3.2.1
jaraco.text==3.5.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
more-itertools==8.7.0
mypy-extensions==0.4.3
packaging==20.9
pluggy==0.13.1
portend==2.7.1
psycopg2==2.8.6
py==1.10.0
pycodestyle==2.6.0
pyflakes==2.2.0
PyJWT==2.0.1
pyparsing==2.4.7
pytest==6.2.2
pytest-freezegun==0.4.2
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
tempora==4.0.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
wrapt==1.12.1
zc.lockfile==2.0
